Turn on sourcemaps in webpack config; regenerate bundle
Because there is a known (by me) bug relating to sourcemaps with webpack
in Rails, I am adding the configuration for using sourcemaps to this
example. "devtool" is a top level key in the webpack.config file, and I
found that warnings in FF were squashed when I configured the sourcemaps
to use the '#' prefix.

In the course of testing this config, I regenerated the bundle.

[Webpack documentation on sourcemap configuration][webpack-sourcemaps]

[wepack-sourcemaps]: http://webpack.github.io/docs/configuration.html#devtool